+++ src/main/java/com/programmers/smrtstore/domain/user/presentation/controller/UserFacade.java
@@ -1,4 +1,4 @@
-package com.programmers.smrtstore.domain.user.application;
+package com.programmers.smrtstore.domain.user.presentation.controller;
 
 import com.programmers.smrtstore.domain.user.presentation.dto.req.DetailShippingRequest;
 import com.programmers.smrtstore.domain.user.presentation.dto.req.UpdateShippingRequest;
